Added more translations

+++ library/Admin/Controller.php
@@ -6,7 +6,7 @@ Class Admin_Controller extends Controller
     public function index()
     {
         if (!User::isAdmin()) {
-            return $this->displayMessage('Keine Berechtigung', false);
+            return $this->displayMessage(_('No Authorization'), false);
         }
 
         $this->view = 'Admin_View_Desktop';
@@ -16,19 +16,22 @@ Class Admin_Controller extends Controller
 
     public function add()
     {
+	    if (!User::isAdmin()) {
+            return $this->displayMessage(_('No Authorization'), false);
+        }
         $zoom = $this->getGetParameter('zoom', FILTER_SANITIZE_NUMBER_INT);
         $lat = $this->getGetParameter('lat', FILTER_SANITIZE_NUMBER_FLOAT);
         $lon = $this->getGetParameter('lon', FILTER_SANITIZE_NUMBER_FLOAT);
         $name = $this->getGetParameter('name');
         if (!$zoom || !$lat || !$lon || !$name) {
-            return $this->displayMessage('Fehlerhafte Eingabe.', false);
+            return $this->displayMessage(_('Invalid input'), false);
         }
 
         if (!System::query("INSERT INTO ".System::getConfig('tbl_prefix')."regions (category, lat, lon, zoom) VALUES(?, ?, ?, ?)", array($name, $lat, $lon, $zoom))) {
-            return $this->displayMessage('Kategorie konnte nicht hinzugefügt werden.', false);
+            return $this->displayMessage(_('Could not add category'), false);
         }
 
-        return $this->displayMessage("'$name' hinzugefügt.", true, array(
+		return $this->displayMessage(_f('%s$1 added', $name), true, array(
                 'id' => System::lastInsertId(),
                 'name' => $name,
                 'lat' => $lat,
@@ -39,20 +42,23 @@ Class Admin_Controller extends Controller
 
     public function drop()
     {
+	    if (!User::isAdmin()) {
+            return $this->displayMessage(_('No Authorization'), false);
+        }
         $id = $this->getGetParameter('id', FILTER_SANITIZE_NUMBER_INT);
         if (!$id) {
-            return $this->displayMessage('Fehlerhafte Eingabe.', false);
+			return $this->displayMessage(_('Invalid input'), false);
         }
 
         if (!System::query("DELETE FROM ".System::getConfig('tbl_prefix')."regions WHERE id = ?", array($id))) {
-            return $this->displayMessage('Kategorie konnte nicht gelöscht werden.', false);
+			return $this->displayMessage(_('Could not delete category'), false);
         }
 
-        return $this->displayMessage('Kategorie gelöscht.', true);
+		return $this->displayMessage(_('Category deleted'), true);
     }
 
     public function getCategories()
-    {
+    {	    
         return $this->categories;
     }
 }
